Remove proof_gen crate
