--- proof_gen/Cargo.toml
@@ -1,36 +1,0 @@
-[package]
-name = "proof_gen"
-description = "Generates block proofs from zero proof IR."
-version = "0.4.0"
-authors = ["Polygon Zero <bgluth@polygon.technology>"]
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-homepage.workspace = true
-keywords.workspace = true
-
-[dependencies]
-log = { workspace = true }
-paste = { workspace = true }
-plonky2 = { workspace = true }
-serde = { workspace = true }
-hashbrown = { workspace = true }
-
-# Local dependencies
-evm_arithmetization = { workspace = true }
-zk_evm_common = { workspace = true }
-
-[features]
-default = ["eth_mainnet"]
-eth_mainnet = [
-    "evm_arithmetization/eth_mainnet"
-]
-cdk_erigon = [
-    "evm_arithmetization/cdk_erigon"
-]
-polygon_pos = [
-    "evm_arithmetization/polygon_pos"
-]
-
-[lints]
-workspace = true
