+++ zero/src/bin/verifier.rs
@@ -5,9 +5,9 @@ use std::fs::File;
 use anyhow::Result;
 use clap::Parser;
 use dotenvy::dotenv;
-use proof_gen::proof_types::GeneratedBlockProof;
 use serde_json::Deserializer;
 use tracing::info;
+use zero::proof_types::GeneratedBlockProof;
 use zero::prover_state::persistence::set_circuit_cache_dir_env_if_not_set;
 
 use self::verifier::*;
