+++ zero/src/lib.rs
@@ -7,6 +7,7 @@ pub mod fs;
 pub mod ops;
 pub mod parsing;
 pub mod pre_checks;
+pub mod proof_types;
 pub mod prover;
 pub mod prover_state;
 pub mod provider;
